backend/requirements.txt
fastapi
uvicorn
python-multipart
groq
pdfplumber
python-dotenv
Pillow
